Decrease linter's memory usage
- Limit linter runner concurrency to 1
- Set GOGC to 50% for both lint phases
- Set linter runner timeout to 90min
